Do not allow methods to be called on invalid VGs and LVs
If an object has .remove called on it, we cannot use the LVM API any more
but the Python object still exists. Add VG_VALID and LV_VALID macros and
use in all method calls to ensure we throw an exception in this case.

Not done for PVs since they cannot be removed via liblvm as yet.

Signed-off-by: Andy Grover <agrover@redhat.com>