remove liblvm object
Instead of requiring users to create a liblvm object, and then calling
methods on it, the module acquires a liblvm handle as part of
initialization. This makes it impossible to instantiate a liblvm object
with a different systemdir, but there is an alternate envvar method for
that obscure use case.

Also, fix whitespace and Yoda conditionals.

Signed-off-by: Andy Grover <agrover@redhat.com>